requirements.txt
apex==0.9.10dev
bbox_visualizer==0.1.0
einops==0.7.0
h5py==3.9.0
hdf5plugin==4.1.3
hydra-core==1.3.2
numpy==1.22.4
omegaconf==2.3.0
opencv_python==4.8.0.76
packaging==23.2
pycocotools==2.0.6
pytorch_lightning==1.8.6
spikingjelly==0.0.0.0.14
StrEnum==0.4.15
torch==2.0.1+cu118
torchdata==0.6.1
torchvision==0.15.2+cu118
tqdm==4.65.0
